+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/service/naming/SiblingNumberGenerator.java
@@ -11,9 +11,19 @@ public interface SiblingNumberGenerator {
    * be appended to the end of the partialAlias to form the complete alias
    * 
    * @param partialAlias the alias being generated, missing only a siblingNumber
-   * @return the lowest useable siblingNumber where the alias formed by {@code (partialAlias + siblingNumber)} does not yet exist
+   * @return the next available siblingNumber that is higher than any existing siblingNumbers
    * @throws IOException
    */
   public <T extends Aliasable> int getNextSiblingNumber(Class<T> clazz, String partialAlias) throws IOException;
+
+  /**
+   * Determines the next available siblingNumber to use for the provided partial alias. Assumes that siblingNumber will
+   * be appended to the end of the partialAlias to form the complete alias
+   * 
+   * @param partialAlias the alias being generated, missing only a siblingNumber
+   * @return the lowest usable siblingNumber where the alias formed by {@code (partialAlias + siblingNumber)} does not yet exist
+   * @throws IOException
+   */
+  public <T extends Aliasable> int getFirstAvailableSiblingNumber(Class<T> clazz, String partialAlias) throws IOException;
 
 }